Put everything related to counting function calls into a different file
